[fixed #159891211]
